\name{Data_Expectativas}
\alias{Data_Expectativas}
\docType{data}
\title{Expectativas Dataset}

\description{
A dataset containing responses to an expectations scale. This dataset is
provided as an example for demonstrating the EFA optimization functions
in the OptimalFactor package.
}

\usage{data(Data_Expectativas)}

\format{
A data frame with observations and item responses for the expectations scale.
}

\source{
Simulated data for demonstration purposes.
}

\examples{
data(Data_Expectativas)
head(Data_Expectativas)
}

\keyword{datasets}
