\name{Data_Personality}
\alias{Data_Personality}
\docType{data}
\title{Personality Dataset}

\description{
A dataset containing responses to a personality assessment scale. This
dataset is provided as an example for demonstrating the EFA optimization
functions in the OptimalFactor package.
}

\usage{data(Data_Personality)}

\format{
A data frame with observations and item responses for personality traits.
}

\source{
Simulated data for demonstration purposes.
}

\examples{
data(Data_Personality)
head(Data_Personality)
}

\keyword{datasets}
