% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_boosting.R
\name{print.discriminant_boosting}
\alias{print.discriminant_boosting}
\title{Print method for discriminant_boosting}
\usage{
\method{print}{discriminant_boosting}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{discriminant_boosting}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}; called for its side effects.
}
\description{
Print method for discriminant_boosting
}
