% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_fit_search.R
\name{print.local_fit_search}
\alias{print.local_fit_search}
\title{Print method for local_fit_search}
\usage{
\method{print}{local_fit_search}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{local_fit_search}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}; called for its side effects.
}
\description{
Print method for local_fit_search
}
