% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification_search.R
\name{print.specification_search}
\alias{print.specification_search}
\title{Print method for specification_search}
\usage{
\method{print}{specification_search}(x, top = 10, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{specification_search}}.}

\item{top}{Number of rows to print. Default 10.}

\item{...}{Ignored.}
}
\description{
Print method for specification_search
}
