% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cfa_boosting.R
\name{export_cfa_boosting}
\alias{export_cfa_boosting}
\title{Export CFA Boosting results to structured data frames}
\usage{
export_cfa_boosting(result)
}
\arguments{
\item{result}{Output from \code{cfa_boosting()}}
}
\value{
A named list of data frames
}
\description{
Export CFA Boosting results to structured data frames
}
